labman-v2/backend/pyproject.toml
[project]
name = "labman-api"
version = "2.0.0"
description = "LabMan API - Academic Lab Management System"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "tomli>=2.0.1",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "app"
module-root = ""
